+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - replaced 'openpyxl' with 'xlsxwriter' to reduce dependencies
+- use 'rich.print' for console output
 
 ## 0.8.0 (2025-04-19)